+++ starboard/extension/experimental/README.md
@@ -2,5 +2,6 @@
 
 Please **DO NOT** use the features in this folder.
 
-Features contained in this folder are strictly experimental and can be changed
-or removed at any time, even within the same Cobalt/Starboard version.
+Features contained in this folder are strictly experimental. They do not
+constitute a stable interface or ABI, and are subject to change or removal
+at any time, even within the same Cobalt/Starboard version.
