+++ starboard/extension/extension_test.cc
@@ -17,8 +17,8 @@
 #include "starboard/extension/accessibility.h"
 #include "starboard/extension/configuration.h"
 #include "starboard/extension/crash_handler.h"
+#include "starboard/extension/experimental/experimental_features.h"
 #include "starboard/extension/experimental/media_buffer_pool.h"
-#include "starboard/extension/experimental_features.h"
 #include "starboard/extension/features.h"
 #include "starboard/extension/font.h"
 #include "starboard/extension/free_space.h"
